Git functions
A git pull function that force imports from your origin. 

And a git push function that automatically adds the time and date as the commit message. It allows you to input 1 optional parameter which is printed behind the time and date in your commit message (as you would with a regular commit message).